--- __tests__/main.test.ts
@@ -1,9 +1,0 @@
-import {getDownloadUrl} from '../src/minikube'
-import * as process from 'process'
-import * as cp from 'child_process'
-import * as path from 'path'
-
-// shows how the runner will run a javascript action with env / stdout protocol
-test('test runs', () => {
-  getDownloadUrl('latest')
-})
