+++ src/main.ts
@@ -1,9 +1,10 @@
 import {getInput, setFailed} from '@actions/core'
 
-import {downloadMinikube, startMinikube} from './minikube'
+import {downloadMinikube} from './download'
+import {startMinikube} from './start'
 
 // main thing :)
-async function run(): Promise<void> {
+const run = async (): Promise<void> => {
   try {
     let minikubeVersion = getInput('minikube-version').toLowerCase()
     minikubeVersion = minikubeVersion === 'stable' ? 'latest' : minikubeVersion
